+++ add.php
@@ -1,31 +1,28 @@
 <?php
+const TAB_TEXT = 'text';
+const TAB_LINK = 'link';
+const TAB_VIDEO = 'video';
+const TAB_QUOTE = 'quote';
+const TAB_PHOTO = 'photo';
 
-date_default_timezone_set("Europe/Moscow");
-
-ini_set('display_errors', 1);
-ini_set('display_startup_errors', 1);
-error_reporting(E_ALL);
-
-$is_auth = rand(0, 1);
-
-$user_name = 'Иван'; // укажите здесь ваше имя
-
-require_once 'functions/main.php';
-require_once 'functions/db.php';
-require_once 'functions/validation_post.php';
-
-$config = require 'config.php';
-$connection = connectDb($config['db']);
+require 'bootstrap.php';
+require 'functions/validators/post/common.php';
+require 'functions/validators/post/text.php';
+require 'functions/validators/post/video.php';
+require 'functions/validators/post/photo.php';
+require 'functions/validators/post/link.php';
+require 'functions/validators/post/quote.php';
+require 'functions/tags.php';
 
 $types = get_types($connection);
-$add_post = $_GET['tab'] ?? '';
+$tab = $_GET['tab'] ?? null;
 
-if (empty($add_post)) {
+if (empty($tab)) {
     header("Location: add.php?tab=text");
     exit();
 }
 
-if (!get_tab($add_post)) {
+if (!get_tab($tab)) {
     header("HTTP/1.0 404 Not Found");
     exit();
 }
@@ -35,7 +32,7 @@ $post_data = [];
 $errors = [];
 $block_errors = null;
 
-$name_type = get_name_type($add_post);
+$name_type = get_name_type($tab);
 $type_id = get_type_id($name_type, $types);
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
@@ -43,18 +40,34 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $post_data = $_POST;
     $file_data = $_FILES;
 
-    $errors = validate_post($post_data, $add_post, $file_data);
+    switch ($tab) {
+        case TAB_TEXT:
+            $errors = validate_post_text($post_data);
+            break;
+        case TAB_LINK:
+            $errors = validate_post_link($post_data);
+            break;
+        case TAB_VIDEO:
+            $errors = validate_post_video($post_data);
+            break;
+        case TAB_QUOTE:
+            $errors = validate_post_quote($post_data);
+            break;
+        case TAB_PHOTO:
+            $errors = validate_post_photo($post_data, $file_data);
+            break;
+    }
 
     if (empty($errors)) {
         if ($file_data) {
             $post_data['img'] = upload_img($file_data['img']);
         }
-
-        $explode_tags = $post_data['tags'] ?? '';
-        if ($explode_tags) {
-            $tags_arr = explode(" ", $explode_tags);
-        }
         $post_id = add_post($connection, $post_data, $type_id);
+
+        $string_tags = $post_data['tags'];
+        if($string_tags) {
+            add_tags($connection, $string_tags, $post_id);
+        }
         if ($post_id) {
             header("Location: post.php?id=" . $post_id);
             exit();
@@ -72,7 +85,7 @@ $title_post = include_template('add_post_title.php', ['errors' => $errors, 'post
 $tags_post = include_template('add_post_tag.php', ['errors' => $errors, 'post_data' => $post_data]);
 $send_form = include_template('add_post_footer.php');
 $page_content = include_template('add.php', [
-    'add_post' => $add_post,
+    'tab' => $tab,
     'types' => $types,
     'block_errors' => $block_errors,
     'send_form' => $send_form,
