+++ functions/db/users.php
@@ -1,0 +1,2 @@
+<?php
+
